Add pile to entry point only
